backend/requirements.txt
fastapi
uvicorn
pydantic-settings
asyncpg
psycopg2-binary
SQLAlchemy[asyncio]
alembic
httpx
pytz
pytest
